BackEnd/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary

# Environment variables
python-dotenv
pydantic
pydantic-settings

# Firebase token verification
google-auth

# Security + JWT (kept for compatibility)
python-jose
passlib[bcrypt]

# pip install -r requirements.txt